updated scripts to be platform aware

+++ precise/build.sh
@@ -3,14 +3,21 @@
 packages=( */debian )
 failure=0
 build_dir="/tmp/cu-cs-pkg-build/"
+platform="precise"
 
 rm -rf "${build_dir}"
 git checkout-index -a -f --prefix="${build_dir}"
-pushd "${build_dir}"
+
+pushd "${build_dir}"/"${platform}"
+if [ $? -ne 0 ]
+then
+    echo "Failed to open ${build_dir}/${platform}"
+    exit 1
+fi
 
 for p in "${packages[@]}"
 do
-    pushd "${p}" > /dev/null
+    pushd "${p}"
     rm -f *.deb
     rm -f *.dsc
     echo "Building ${p}..."
@@ -22,7 +29,7 @@ do
     else
 	echo "Build succeeded"
     fi
-    popd > /dev/null
+    popd
 done
 
 if [ ${failure} -ne 0 ]
